Improved staff lis view
