+++ src/components/fragments/util/Button.tsx
@@ -1,5 +1,4 @@
-import { MouseEventHandler } from "react";
-import { ReactChild, ReactChildren } from "react";
+import { MouseEventHandler, ReactChild, ReactChildren } from "react";
 import { useDarkMode } from "../../../hooks/useDarkMode";
 import { Button as Inner } from "react-bootstrap";
